# Conflicts:
#	deno_webgpu/lib.rs
#	examples/mipmap/src/main.rs
#	tests/tests/shader_primitive_index/mod.rs

+++ wgpu-core/src/command/draw.rs
@@ -66,7 +66,7 @@ pub enum RenderCommandError {
     #[error("Render bundle {0:?} is invalid")]
     InvalidRenderBundle(id::RenderBundleId),
     #[error("Bind group index {index} is greater than the device's requested `max_bind_group` limit {max}")]
-    BindGroupIndexOutOfRange { index: u8, max: u32 },
+    BindGroupIndexOutOfRange { index: u32, max: u32 },
     #[error("Dynamic buffer offset {0} does not respect device's requested `{1}` limit {2}")]
     UnalignedBufferOffset(u64, &'static str, u32),
     #[error("Number of buffer offsets ({actual}) does not match the number of dynamic bindings ({expected})")]
@@ -148,7 +148,7 @@ pub struct Rect<T> {
 )]
 pub enum RenderCommand {
     SetBindGroup {
-        index: u8,
+        index: u32,
         num_dynamic_offsets: u8,
         bind_group_id: id::BindGroupId,
     },
